#python_code_will_remove_later/requirements.txt
numpy
requests
pillow
gradio
opencv-python
websockets
websocket-client